fix objectlookup single item bug
