+++ src/main/resources/test_files/runtime/PostFixObjectLookup/PostFixObjectLookup1.iq
@@ -1,0 +1,5 @@
+(:JIQS: ShouldRun; Output="bar" :)
+({"foo" : "bar"}).foo
+
+(: single object :)
+
