+++ src/main/resources/test_files/runtime/PostFixObjectLookup/PostFixObjectLookup2.iq
@@ -1,0 +1,5 @@
+(:JIQS: ShouldRun; Output="( bar, foobar)" :)
+({"foo" : "bar"}, {"foo" : "foobar"}).foo
+
+(: two objects :)
+
