+++ src/main/resources/test_files/runtime/PostFixObjectLookup/PostFixObjectLookup3.iq
@@ -1,0 +1,5 @@
+(:JIQS: ShouldRun; Output="" :)
+().foo
+
+(: empty sequence :)
+
